+++ poliastro/tests/test_twobody.py
@@ -2,7 +2,8 @@
 import pytest
 
 import numpy as np
-from numpy.testing import assert_almost_equal, assert_array_almost_equal
+from numpy.testing import (assert_equal, assert_almost_equal,
+                           assert_array_almost_equal)
 
 from astropy import units as u
 from astropy import time
@@ -10,12 +11,16 @@ from astropy import time
 from poliastro.bodies import Sun, Earth
 from poliastro.twobody import State
 
+from poliastro.twobody.rv import rv2coe
+from poliastro.twobody.classical import coe2rv, coe2mee
+from poliastro.twobody.equinoctial import mee2coe
+
 
 def test_state_has_attractor_given_in_constructor():
     _d = 1.0 * u.AU  # Unused distance
     _ = 0.5 * u.one  # Unused dimensionless value
     _a = 1.0 * u.deg  # Unused angle
-    ss = State.from_elements(Sun, _d, _, _a, _a, _a, _a)
+    ss = State.from_classical(Sun, _d, _, _a, _a, _a, _a)
     assert ss.attractor == Sun
 
 
@@ -25,7 +30,7 @@ def test_default_time_for_new_state():
     _a = 1.0 * u.deg  # Unused angle
     _body = Sun  # Unused body
     expected_epoch = time.Time("J2000", scale='utc')
-    ss = State.from_elements(_body, _d, _, _a, _a, _a, _a)
+    ss = State.from_classical(_body, _d, _, _a, _a, _a, _a)
     assert ss.epoch == expected_epoch
 
 
@@ -33,22 +38,24 @@ def test_state_has_elements_given_in_constructor():
     # Mars data from HORIZONS at J2000
     a = 1.523679 * u.AU
     ecc = 0.093315 * u.one
+    p = a * (1 - ecc**2)
     inc = 1.85 * u.deg
     raan = 49.562 * u.deg
     argp = 286.537 * u.deg
     nu = 23.33 * u.deg
-    ss = State.from_elements(Sun, a, ecc, inc, raan, argp, nu)
-    assert ss.elements == (a, ecc, inc, raan, argp, nu)
+    ss = State.from_classical(Sun, p, ecc, inc, raan, argp, nu)
+    assert ss.coe == (p, ecc, inc, raan, argp, nu)
 
 
 def test_state_has_individual_elements():
     a = 1.523679 * u.AU
     ecc = 0.093315 * u.one
+    p = a * (1 - ecc**2)
     inc = 1.85 * u.deg
     raan = 49.562 * u.deg
     argp = 286.537 * u.deg
     nu = 23.33 * u.deg
-    ss = State.from_elements(Sun, a, ecc, inc, raan, argp, nu)
+    ss = State.from_classical(Sun, p, ecc, inc, raan, argp, nu)
     assert ss.a == a
     assert ss.ecc == ecc
     assert ss.inc == inc
@@ -63,7 +70,7 @@ def test_state_raises_unitserror_if_elements_units_are_wrong():
     _a = 1.0 * u.deg  # Unused angle
     wrong_angle = 1.0 * u.AU
     with pytest.raises(u.UnitsError) as excinfo:
-        ss = State.from_elements(Sun, _d, _, _a, _a, _a, wrong_angle)
+        ss = State.from_classical(Sun, _d, _, _a, _a, _a, wrong_angle)
     assert ("UnitsError: Units must be consistent"
             in excinfo.exconly())
 
@@ -98,18 +105,6 @@ def test_geosync_has_proper_period():
     assert_almost_equal(ss.period.to(u.min).value, expected_period, decimal=0)
 
 
-def test_parabolic_elements_fail_early():
-    attractor = Earth
-    ecc = 1.0 * u.one
-    _d = 1.0 * u.AU  # Unused distance
-    _ = 0.5 * u.one  # Unused dimensionless value
-    _a = 1.0 * u.deg  # Unused angle
-    with pytest.raises(ValueError) as excinfo:
-        ss = State.from_elements(attractor, _d, ecc, _a, _a, _a, _a)
-    assert ("ValueError: For parabolic orbits use State.parabolic instead"
-            in excinfo.exconly())
-
-
 def test_parabolic_has_proper_eccentricity():
     attractor = Earth
     _d = 1.0 * u.AU  # Unused distance
@@ -117,7 +112,17 @@ def test_parabolic_has_proper_eccentricity():
     _a = 1.0 * u.deg  # Unused angle
     expected_ecc = 1.0 * u.one
     ss = State.parabolic(attractor, _d, _a, _a, _a, _a)
-    assert_almost_equal(ss.ecc, 1.0 * u.one)
+    assert_almost_equal(ss.ecc, expected_ecc)
+
+
+def test_parabolic_has_infinite_semimajor_axis():
+    attractor = Earth
+    _d = 1.0 * u.AU  # Unused distance
+    _ = 0.5 * u.one  # Unused dimensionless value
+    _a = 1.0 * u.deg  # Unused angle
+    expected_a = np.inf * u.km
+    ss = State.parabolic(attractor, _d, _a, _a, _a, _a)
+    assert_equal(ss.a, expected_a)
 
 
 def test_parabolic_has_zero_energy():
@@ -126,8 +131,7 @@ def test_parabolic_has_zero_energy():
     _ = 0.5 * u.one  # Unused dimensionless value
     _a = 1.0 * u.deg  # Unused angle
     ss = State.parabolic(attractor, _d, _a, _a, _a, _a)
-    E = ss.v.dot(ss.v) / 2 - attractor.k / np.sqrt(ss.r.dot(ss.r))
-    assert_almost_equal(E.value, 0.0)
+    assert_almost_equal(ss.energy.value, 0.0)
 
 
 def test_perigee_and_apogee():
@@ -135,8 +139,9 @@ def test_perigee_and_apogee():
     expected_r_p = 300 * u.km
     a = (expected_r_a + expected_r_p) / 2
     ecc = expected_r_a / a - 1
+    p = a * (1 - ecc**2)
     _a = 1.0 * u.deg  # Unused angle
-    ss = State.from_elements(Earth, a, ecc, _a, _a, _a, _a)
+    ss = State.from_classical(Earth, p, ecc, _a, _a, _a, _a)
     assert_almost_equal(ss.r_a.to(u.km).value,
                         expected_r_a.to(u.km).value)
     assert_almost_equal(ss.r_p.to(u.km).value,
@@ -148,14 +153,13 @@ def test_convert_from_rv_to_coe():
     attractor = Earth
     p = 11067.790 * u.km
     ecc = 0.83285 * u.one
-    a = p / (1 - ecc ** 2)
     inc = 87.87 * u.deg
     raan = 227.89 * u.deg
     argp = 53.38 * u.deg
     nu = 92.335 * u.deg
     expected_r = [6525.344, 6861.535, 6449.125]  # km
     expected_v = [4.902276, 5.533124, -1.975709]  # km / s
-    r, v = State.from_elements(Earth, a, ecc, inc, raan, argp, nu).rv()
+    r, v = State.from_classical(Earth, p, ecc, inc, raan, argp, nu).rv()
     assert_array_almost_equal(r.value, expected_r, decimal=1)
     assert_array_almost_equal(v.value, expected_v, decimal=5)
 
@@ -166,21 +170,51 @@ def test_convert_from_coe_to_rv():
     r = [6524.384, 6862.875, 6448.296] * u.km
     v = [4.901327, 5.533756, -1.976341] * u.km / u.s
     ss = State.from_vectors(Earth, r, v)
-    a, ecc, inc, raan, argp, nu = ss.elements
-    p = ss.p
-    assert_almost_equal(p.value, 11067.79, decimal=0)
+    p, ecc, inc, raan, argp, nu = ss.coe
+    assert_almost_equal(p.to(u.km).value, 11067.79, decimal=0)
     assert_almost_equal(ecc.value, 0.832853, decimal=4)
-    assert_almost_equal(inc.value, 87.870, decimal=2)
-    assert_almost_equal(raan.value, 227.89, decimal=1)
-    assert_almost_equal(argp.value, 53.38, decimal=2)
-    assert_almost_equal(nu.value, 92.335, decimal=2)
+    assert_almost_equal(inc.to(u.deg).value, 87.870, decimal=2)
+    assert_almost_equal(raan.to(u.deg).value, 227.89, decimal=1)
+    assert_almost_equal(argp.to(u.deg).value, 53.38, decimal=2)
+    assert_almost_equal(nu.to(u.deg).value, 92.335, decimal=2)
+
+
+def test_convert_between_coe_and_rv_is_transitive():
+    k = Earth.k.to(u.km**3 / u.s**2).value  # u.km**3 / u.s**2
+    p = 11067.790  # u.km
+    ecc = 0.83285  # u.one
+    inc = np.deg2rad(87.87)  # u.rad
+    raan = np.deg2rad(227.89)  # u.rad
+    argp = np.deg2rad(53.38)  # u.rad
+    nu = np.deg2rad(92.335)  # u.rad
+
+    expected_res = (p, ecc, inc, raan, argp, nu)
+
+    res = rv2coe(k, *coe2rv(k, *expected_res))
+
+    assert_array_almost_equal(res, expected_res)
+
+
+def test_convert_between_coe_and_mee_is_transitive():
+    p = 11067.790  # u.km
+    ecc = 0.83285  # u.one
+    inc = np.deg2rad(87.87)  # u.rad
+    raan = np.deg2rad(227.89)  # u.rad
+    argp = np.deg2rad(53.38)  # u.rad
+    nu = np.deg2rad(92.335)  # u.rad
+
+    expected_res = (p, ecc, inc, raan, argp, nu)
+
+    res = mee2coe(*coe2mee(*expected_res))
+
+    assert_array_almost_equal(res, expected_res)
 
 
 def test_apply_zero_maneuver_returns_equal_state():
     _d = 1.0 * u.AU  # Unused distance
     _ = 0.5 * u.one  # Unused dimensionless value
     _a = 1.0 * u.deg  # Unused angle
-    ss = State.from_elements(Sun, _d, _, _a, _a, _a, _a)
+    ss = State.from_classical(Sun, _d, _, _a, _a, _a, _a)
     dt = 0 * u.s
     dv = [0, 0, 0] * u.km / u.s
     ss_new = ss.apply_maneuver([(dt, dv)])
@@ -194,7 +228,7 @@ def test_apply_maneuver_changes_epoch():
     _d = 1.0 * u.AU  # Unused distance
     _ = 0.5 * u.one  # Unused dimensionless value
     _a = 1.0 * u.deg  # Unused angle
-    ss = State.from_elements(Sun, _d, _, _a, _a, _a, _a)
+    ss = State.from_classical(Sun, _d, _, _a, _a, _a, _a)
     dt = 1 * u.h
     dv = [0, 0, 0] * u.km / u.s
     ss_new = ss.apply_maneuver([(dt, dv)])
@@ -205,7 +239,7 @@ def test_perifocal_points_to_perigee():
     _d = 1.0 * u.AU  # Unused distance
     _ = 0.5 * u.one  # Unused dimensionless value
     _a = 1.0 * u.deg  # Unused angle
-    ss = State.from_elements(Sun, _d, _, _a, _a, _a, _a)
+    ss = State.from_classical(Sun, _d, _, _a, _a, _a, _a)
     p, _, _ = ss.pqw()
     assert_almost_equal(p, ss.e_vec / ss.ecc)